Replace unsigned int cast with a suffix

+++ src/kvirc/kernel/KviApplication.cpp
@@ -850,7 +850,7 @@ void KviApplication::notifierMessage(KviWindow * pWnd, int iIconId, const QStrin
 			// org.freedesktop.Notifications.Notify
 			QVariantList args;
 			args << QString("KVIrc");                          // application name
-			args << QVariant((unsigned int) 0);                // notification id
+			args << QVariant(0u);                              // notification id
 			args << szIcon;                                    // application icon
 			args << szTitle;                                   // summary text
 			args << szText;                                    // detailed text
